Exhuastively search for minimum sample rate
Available ample rates are not necessarily sorted, so we need to check
all of them to find the best. We sort this by USB bandwidth. Currently
this will hit Saleae several times to ask for the connected devices and
active channels, but I don't think that's a big deal in practice and
makes `get_bandwidth` more accessible for other uses.

Closes #8.