Add parameter to launch_logic allowing the caller to specify path
This commit updates `launch_logic` with a `logic_path` parameter that
allows the caller to specify the path to the Logic application. This
allows users with Logic installed to a non-standard location to still be
able to use this method.

If this option is provided, it will be the preferred executable and we
will not search for Logic in other possible locations (i.e. if it isn't
there, this will raise an exception). If this option is not provided,
standard install locations will be attempted.